refactor(exoplayer): refactor exoplayer to conform to new playback interface
