+++ app/src/main/kotlin/com/globo/clappr/app/PlayerActivity.kt
@@ -11,7 +11,6 @@ import com.globo.clappr.base.BaseObject
 import com.globo.clappr.base.Options
 import com.globo.clappr.playback.ExoPlayerPlayback
 import com.globo.clappr.plugin.Loader
-import com.globo.clappr.plugin.Playback.ExoPlayerPlayBack
 
 class PlayerActivity : Activity() {
 
@@ -32,7 +31,7 @@ class PlayerActivity : Activity() {
     override fun onResume() {
         super.onResume()
         val viewGroup = findViewById(R.id.player) as FrameLayout
-        ExoPlayerPlayBack.containerView = viewGroup
+        ExoPlayerPlayback.containerView = viewGroup
         val duration = player?.core?.activeContainer?.playback?.duration
 
         player?.attachTo(viewGroup)
