+++ build.gradle
@@ -1,8 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    def buildToolsVersion = '23.0.3'
-    def kotlinVersion = '1.0.4'
+    ext.kotlin_version = '1.0.4'
 
     repositories {
         mavenCentral()
@@ -10,10 +9,11 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
-        classpath 'com.android.tools.build:gradle:2.2.0'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath 'com.android.tools.build:gradle:2.2.2'
         classpath 'net.saliman:gradle-cobertura-plugin:2.2.5'
         classpath 'org.kt3k.gradle.plugin:coveralls-gradle-plugin:2.3.1'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
